Remove old agent watcher and agent installer that is no longer in use.
The agent-watcher was replaced some time ago by a watcher that instead
watchers the telepresence-agents config-map. The old installer that
predates 2.6.0 is never used unless the traffic-manager predates 2.6.0,
and we don't have backward compatibility tests that go beyond that.
