+++ pkg/install/constants.go
@@ -2,14 +2,13 @@ package install
 
 const (
 	AgentContainerName        = "traffic-agent"
+	InitContainerName         = "tel-agent-init"
 	AgentAnnotationVolumeName = "traffic-annotations"
 	AgentInjectorName         = "agent-injector"
 	DomainPrefix              = "telepresence.getambassador.io/"
 	InjectAnnotation          = DomainPrefix + "inject-" + AgentContainerName
-	ServicePortAnnotation     = DomainPrefix + "inject-service-port"
 	ServiceNameAnnotation     = DomainPrefix + "inject-service-name"
 	ManualInjectAnnotation    = DomainPrefix + "manually-injected"
 	ManagerAppName            = "traffic-manager"
 	MutatorWebhookTLSName     = "mutator-webhook-tls"
-	TelAppMountPoint          = "/tel_app_mounts"
 )
